requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
pymysql==1.0.2
jieba==0.42.1
pandas==1.3.3
numpy==1.21.2
python-dotenv==0.19.0
flask-cors==3.0.10 